Moved my comparator to hudson/plugins/report/jck/main/comparator
