+++ src/main/java/hudson/plugins/report/jck/main/comparator/QueryString.java
@@ -1,4 +1,4 @@
-package hudson.plugins.report.jck.comparer;
+package hudson.plugins.report.jck.main.comparator;
 
 import java.io.File;
 import java.util.ArrayList;
